move resolve-packages to own feature
