+++ Cargo.toml
@@ -8,23 +8,22 @@ repository = "https://github.com/Relacibo/typst-as-lib"
 readme = "README.md"
 keywords = ["template", "typst"]
 categories = ["template-engine"]
-exclude = [
-    "examples/*",
-]
+exclude = ["examples/*"]
+
+[features]
+default = ["packages"]
+packages = ["dep:binstall-tar", "dep:flate2", "dep:ureq"]
 
 [dependencies]
-binstall-tar = "0.4.42"
+binstall-tar = { version = "0.4.42", optional = true }
 chrono = "0.4.38"
 comemo = "0.4.0"
 ecow = "0.2.2"
-flate2 = "1.0.30"
+flate2 = { version = "1.0.30", optional = true }
 thiserror = "1.0.62"
 typst = "0.11.1"
-ureq = "2.10.0"
+ureq = { version = "2.10.0", optional = true }
 
 [dev-dependencies]
 derive_typst_intoval = "0.2.1"
-tar = "0.4.41"
 typst-pdf = "0.11.1"
-ureq = "2.10.0"
-zune-inflate = "0.2.54"
